requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
distlib==0.3.0
Django==3.0.3
django-environ==0.4.5
djangorestframework==3.11.0
filelock==3.0.12
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.1
pbr==5.4.4
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
stevedore==1.32.0
virtualenv==20.0.5
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.8
wrapt==1.11.2
